refactor: jdbctemplate에 통합
