+++ jdbc/src/main/java/nextstep/jdbc/PreparedStatementSetter.java
@@ -1,0 +1,9 @@
+package nextstep.jdbc;
+
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+
+public interface PreparedStatementSetter {
+
+    void setValues(PreparedStatement pstmt) throws SQLException;
+}
