+++ jdbc/src/main/java/nextstep/jdbc/RowMapper.java
@@ -1,0 +1,9 @@
+package nextstep.jdbc;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+public interface RowMapper {
+
+    Object mapRow(ResultSet rs) throws SQLException;
+}
